[KYUUBI #6554] Delete redundant code related to zorder
