+++ extensions/spark/kyuubi-extension-spark-3-3/src/test/scala/org/apache/spark/sql/ZorderSuite.scala
@@ -24,7 +24,7 @@ import org.apache.spark.sql.internal.SQLConf
 import org.apache.kyuubi.sql.{KyuubiSQLConf, SparkKyuubiSparkSQLParser}
 import org.apache.kyuubi.sql.zorder.Zorder
 
-trait ZorderSuiteSpark33 extends ZorderSuiteBase {
+trait ZorderSuiteSpark extends ZorderSuiteBase {
 
   test("Add rebalance before zorder") {
     Seq("true" -> false, "false" -> true).foreach { case (useOriginalOrdering, zorder) =>
@@ -115,10 +115,10 @@ trait ParserSuite { self: ZorderSuiteBase =>
 
 class ZorderWithCodegenEnabledSuite
   extends ZorderWithCodegenEnabledSuiteBase
-  with ZorderSuiteSpark33
+  with ZorderSuiteSpark
   with ParserSuite {}
 
 class ZorderWithCodegenDisabledSuite
   extends ZorderWithCodegenDisabledSuiteBase
-  with ZorderSuiteSpark33
+  with ZorderSuiteSpark
   with ParserSuite {}
